requirements.txt
langchain
python-dotenv
ipykernel
beautifulsoup4
langchain_community
faiss-cpu
streamlit